isolate data containers from solver
